added decorator project files
