+++ person.rb
@@ -4,7 +4,7 @@ class Person < Nameable
   attr_accessor :name, :age
   attr_reader :id
 
-  def initialize(name, age, parent_permission: true)
+  def initialize(age,name, parent_permission: true)
     super()
     @id = generate_id
     @name = name
@@ -16,8 +16,8 @@ class Person < Nameable
     of_age? || @parent_permission
   end
 
-  def create_name?
-    @name != ''
+  def correct_name
+   p @name 
   end
 
   private
@@ -31,5 +31,33 @@ class Person < Nameable
   end
 end
 
-person = Person.new('John', 17)
-p person
+class BaseDecorator < Nameable
+    def initialize(nameable)
+      super()
+      @nameable = nameable
+    end
+  
+    def correct_name
+      @nameable.correct_name
+    end
+end
+
+class CapitalizeDecorator < BaseDecorator
+    def correct_name
+      @nameable.correct_name.capitalize
+    end
+end
+
+class TrimmerDecorator < BaseDecorator
+    def correct_name
+      @nameable.correct_name[0..9]
+    end
+end
+  
+  
+person = Person.new(22, 'maximilianus')
+person.correct_name
+capitalized_person = CapitalizeDecorator.new(person)
+puts capitalized_person.correct_name
+capitalized_trimmed_person = TrimmerDecorator.new(capitalized_person)
+puts capitalized_trimmed_person.correct_name
