--- timmerdecorator.rb
@@ -1,7 +1,0 @@
-require './basedecorator'
-
-class TrimmerDecorator < BaseDecorator
-  def correct_name
-    @nameable.correct_name[0..9]
-  end
-end
